restaurants/requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
